requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
streamlit 
google-generativeai
python-dotenv
langchain
pyPDF2
faiss-cpu
langchain_google_genai
